+++ plot.h
@@ -8,7 +8,7 @@
 class Plot {
 public: 
     void draw(sf::RenderWindow& window);
-    void add(std::pair<float, float> value);
+    void add(std::pair<double, double> value);
     Plot(unsigned int width, unsigned int height);
     
 private:
@@ -16,9 +16,9 @@ private:
     unsigned int y;
     unsigned int height;
     unsigned int width; // width of plot in pixels
-    float interval; // time interval in seconds
-    std::queue< std::pair<float, float> > values; // queue of points(time, y-value)
-    float maxY; // Used to dynamicly resize the plot window 
+    double interval; // time interval in seconds
+    std::queue< std::pair<double, double> > values; // queue of points(time, y-value)
+    double maxY; // Used to dynamicly resize the plot window 
 };
